requirements.txt
langchain
transformers
torch
sentence-transformers
python-dotenv
pytest
langchain_community
